+++ tools/lib-alert-tree/tests/test_kubernetes.py
@@ -49,6 +49,7 @@ def test_deployment_alerts():
 
 def test_pod_alerts():
     """Check Pod alerts."""
+    # pylint: disable=line-too-long
     test = D.warning(
         "test",
         children=kubernetes.pod_alerts("my-pod", namespace="my-ns"),
@@ -59,6 +60,7 @@ def test_pod_alerts():
         == textwrap.dedent(
             """
             test{severity='warning'}
+            ├── KubePodCrashLooping{namespace=~'my-ns', pod=~'my-pod', severity='warning'}
             └── KubePodNotReady{namespace=~'my-ns', pod=~'my-pod', severity='warning'}
             """
         ).lstrip()
@@ -81,6 +83,25 @@ def test_statefulset_alerts():
             ├── KubeStatefulSetGenerationMismatch{namespace=~'my-ns', severity='warning', statefulset=~'my-sts'}
             ├── KubeStatefulSetReplicasMismatch{namespace=~'my-ns', severity='warning', statefulset=~'my-sts'}
             └── KubeStatefulSetUpdateNotRolledOut{namespace=~'my-ns', severity='warning', statefulset=~'my-sts'}
+            """
+        ).lstrip()
+    )
+
+
+def test_job_alerts():
+    """Check Job alerts."""
+    test = D.warning(
+        "test",
+        children=kubernetes.job_alerts("my-job", namespace="my-ns"),
+        relationship=Relationship.ANY,
+    )
+    assert (
+        test.build_tree().show(stdout=False)
+        == textwrap.dedent(
+            """
+            test{severity='warning'}
+            ├── KubeJobCompletion{job=~'my-job', namespace=~'my-ns', severity='warning'}
+            └── KubeJobFailed{job=~'my-job', namespace=~'my-ns', severity='warning'}
             """
         ).lstrip()
     )